requirements.txt
streamlit
openai>=1.0.0
requests
beautifulsoup4
PyMuPDF
python-docx
fpdf